creation function refactor
