+++ packages/webapp/src/components/LocationCreationModal/index.jsx
@@ -20,9 +20,23 @@ import styles from './styles.module.scss';
 import Button from '../Form/Button';
 import PropTypes from 'prop-types';
 import { Info, Semibold } from '../Typography';
+import { useDispatch, useSelector } from 'react-redux';
+import { userFarmSelector } from '../../containers/userFarmSlice';
+import { setMapFilterSetting } from '../../containers/Map/mapFilterSettingSlice';
+import history from '../../history';
 
-export default function LocationCreationModal({ title, body, dismissModal, locationCreation }) {
+export default function LocationCreationModal({ title, body, dismissModal }) {
+  const dispatch = useDispatch();
   const { t } = useTranslation();
+  const { farm_id } = useSelector(userFarmSelector);
+
+  const onCreateLocation = () => {
+    const payload = {};
+    payload.farm_id = farm_id;
+    payload.addDrawer = true;
+    dispatch(setMapFilterSetting(payload));
+    history.push('/map');
+  };
 
   return (
     <Modal dismissModal={dismissModal}>
@@ -35,7 +49,7 @@ export default function LocationCreationModal({ title, body, dismissModal, locat
           <Button onClick={dismissModal} sm>
             {t('common:BACK')}
           </Button>
-          <Button onClick={locationCreation} sm>
+          <Button onClick={onCreateLocation} sm>
             {t('LOCATION_CREATION.CREATE_BUTTON')}
           </Button>
         </div>
